[PAXWEB-996] Missing synchronization in pax-web-resources-extender
Internal storage will be replaced with Java 5 ConcurrentHashMap and CopyOnWriteArrayList (for the shadowed entries).
Since especially locking any bundles stop-method is not advisable the fix is intended to remove all locking and use standard collections which offer the best trade-off between access and visibility.
It is possible that a resource is requested while the extender-thread is populating the list. Depending on the state of the ConcurrentHashMap, this resource might not yet be available.
This is a small tradeoff compared to the complexity of manual synchronization.