[code cleanup] commons-code removed
Commons-Codec was only used for Base4-decoding which is available in JDK 8 as well.
The dependency has been removed though it is still necessary for some test-cases (mostly JSF) which drag commons-codec in as transitive dependency.

During cleanup the configuration for JSF (MyFaces) was moved to TestConfiguration which is used by some itests.