Code-Analysis plugins moved to profile
The findbugs and especially checkstyle plugins cause trouble in an IDE when executing a maven goal other than the root-pom because the configuration files cannot be found.
The plugins were moved to a dedicated profile which is active by default and can be disabled on IDE-level. This elimates the need to create dedicated run-configuration in the IDE where those plugins are skipped.